pyproject.toml
[tool.poetry]
name = "eagleview"
version = "1.2"
description = "Easy Visualisation of Plots"
authors = ["Aditya"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
matplotlib = "^3.0"
pandas = "^2.0"
pillow = "^10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
